requirements.txt
Flask==2.2.2
openai==0.27.0
whisper==1.1.0
youtube-transcript-api==0.4.4
python-dotenv==1.0.0
pytube==12.1.0
